allow for a test_config parameter to be passed in in the ui
