+++ lib/bindConnectors/index.js
@@ -3,68 +3,21 @@
 * with the node connector sdk. Throws error if validations fail. We don't
 * wanna be starting a server with invalid config.
 */
-var util 					      = require('util');
-var _ 					        = require('lodash');
-var when 						    = require('when');
-var Threadneedle        = require('@trayio/threadneedle');
-var camelCase 			    = require('mout/string/camelCase');
-var bindMessage         = require('./bindMessage');
-var bindDynamicOutput         = require('./bindDynamicOutput');
-var bindTriggerRequest  = require('./bindTriggerRequest');
-var bindTriggerResponse = require('./bindTriggerResponse');
-var formatError 		    = require('./formatError');
-var formatMessage       = require('./formatMessage');
+var util 	 		  = require('util');
+var _ 				  = require('lodash');
+var when 		 	  = require('when');
+var formatError 	  = require('./formatError');
+var formatMessage     = require('./formatMessage');
+var bindMessageHooks  = require('./bindMessageHooks');
+var parseConfig		  = require('../parseConfig');
+
 
 
 module.exports = function (config, options) {
 	// console.log(util.inspect(config, false, null));
 
-	// Internal-only message function handlers. Exactly the same as
-	// falafel[connectorName][methodName](), but only set for use below
-	// in the handler.
-	var messageHooks = {};
+	var messageHooks = bindMessageHooks(config, options);
 
-	// For each connector, bind all the messages. Note that we can't
-	// have a conflict in namings.
-	_.each(config, function (connectorConfig) {
-
-		// console.log('Binding connector:', connectorConfig.name);
-
-		// Declare threadneedle global config for the connector
-		var threadneedle = new Threadneedle();
-		threadneedle.global(connectorConfig.globalModel);
-
-		// Bind each message
-		_.each(connectorConfig.messages || [], function (message) {
-
-			// Add the message into threadneedle
-			messageHooks[message.name] = bindMessage(message, threadneedle, connectorConfig);
-
-			// var possibleSubMessages = _.pick(message, ['model', 'request', 'response', 'destroy']);
-			if (_.isFunction(message.destroy) || _.isObject(message.destroy)) {
-				var destroyMessageName = message.name+'_destroy';
-				messageHooks[destroyMessageName] = bindMessage(_.defaults({
-					name: destroyMessageName,
-					model: message.destroy
-				}, message, options), threadneedle, connectorConfig);
-			}
-
-			if (_.isFunction(message.request) || _.isObject(message.request)) {
-				messageHooks[message.name+'_request'] = bindTriggerRequest(message, options);
-			}
-
-			if (_.isFunction(message.response)) {
-				messageHooks[message.name+'_response'] = bindTriggerResponse(message, options);
-			}
-
-			if (_.isFunction(message.dynamicOutput)) {
-				messageHooks[message.name + '_output_schema'] = bindDynamicOutput(message, options);
-			}
-
-		});
-
-
-	});
 
 	// Return a message handler function. This is what is exported
 	// at the top level.
@@ -77,45 +30,54 @@ module.exports = function (config, options) {
 		var promises = _.map(events, function (event) {
 			return when.promise(function (resolve, reject) {
 
-				var output;
-
 				// Handle error invalid payload
 				if (!event.header || !event.header.message) {
-					resolve(formatError({
+					return resolve(formatError({
 						code: 'invalid_input',
 						message: 'Invalid message received.'
 					}, event));
 				}
 
-				// If it's a trigger,
+
+				var hooks;
+
+				// TEST MODE
+				// If the event header has the `test_config` key passed in, this should be treated 
+				// as a total override, as we're in test mode. In this case, parse the `test_config` 
+				// and re-bind the message hooks
+				if (_.isObject(event.header.test_config)) {
+					var testConfig = parseConfig(event.header.test_config);
+					hooks = bindMessageHooks(testConfig, options);
+				} else {
+					hooks = messageHooks;
+				}
+
 
 				// Handle message not existing
-				else if (!_.isFunction(messageHooks[event.header.message])) {
-					resolve(formatError({
+				if (!_.isFunction(hooks[event.header.message])) {
+					return resolve(formatError({
 						code: 'not_implemented',
 						message: 'Could not find a valid message handler for ' + event.header.message
 					}, event));
 				}
 
-				// All good, let's run
-				else {
+				// Got this far? All good, let's run
 
-					// NOTE: we actually need try catch as when.js
-					// promises catch errors in the reject anyway.
+				// NOTE: we actually need try catch as when.js
+				// promises catch errors in the reject anyway.
 
-					// Run the API call, and respond with the appropriate stuff
-					messageHooks[event.header.message](event)
+				// Run the API call, and respond with the appropriate stuff
+				hooks[event.header.message](event)
 
-					.done(
-						function (response) {
-							resolve(formatMessage(event, response.body, response.version));
-						}, function (err) {
-							resolve(formatError(err, event));
-						}
-					);
+				.done(
+					function (response) {
+						resolve(formatMessage(event, response.body, response.version));
+					}, function (err) {
+						resolve(formatError(err, event));
+					}
+				);
 
 
-				}
 			});
 		});
