+++ lib/index.js
@@ -4,6 +4,7 @@ var bindHelpers         = require('./bindHelpers');
 var getConnectors       = require('./getConnectors');
 var bindConnectors      = require('./bindConnectors');
 var isLegacyConnector 	= require('./utils/isLegacy');
+var parseConfig 		= require('./parseConfig');
 
 
 
@@ -67,8 +68,11 @@ var Falafel = function () {
 			// NEW (ARTISAN)
 			else {
 
-				// Bind the models and make the connector actually work
-				var connectorsConfig = getConnectors(options.directory);
+				// Get the raw config json from the file
+				var rawConfig = require(options.directory+'/config.json');
+
+				// Parse it into a config that Falafel can understand
+				var connectorsConfig = parseConfig(rawConfig);
 
 				// Add in the file upload & download stuff
 				require('./fileHandler')(options);
