+++ lib/parseConfig/getFunctionParameter.js
