+++ lib/parseConfig/getGlobalModel.js
