+++ lib/parseConfig/getModel.js
