+++ lib/parseConfig/normalizeModelParameter.js
